+++ google-cloud-bigquerystorage/src/test/java/com/google/cloud/bigquery/storage/v1/JsonStreamWriterTest.java
@@ -23,11 +23,9 @@ import com.google.api.client.util.Sleeper;
 import com.google.api.core.ApiFuture;
 import com.google.api.gax.batching.FlowControlSettings;
 import com.google.api.gax.batching.FlowController;
-import com.google.api.gax.core.ExecutorProvider;
 import com.google.api.gax.core.InstantiatingExecutorProvider;
 import com.google.api.gax.core.NoCredentialsProvider;
 import com.google.api.gax.grpc.testing.LocalChannelProvider;
-import com.google.api.gax.grpc.testing.MockGrpcService;
 import com.google.api.gax.grpc.testing.MockServiceHelper;
 import com.google.cloud.bigquery.storage.test.JsonTest;
 import com.google.cloud.bigquery.storage.test.SchemaTest;
@@ -36,7 +34,7 @@ import com.google.cloud.bigquery.storage.test.Test.FooType;
 import com.google.cloud.bigquery.storage.test.Test.RepetitionType;
 import com.google.cloud.bigquery.storage.test.Test.UpdatedFooType;
 import com.google.cloud.bigquery.storage.v1.ConnectionWorkerPool.Settings;
-import com.google.cloud.bigquery.storage.v1.Exceptions.AppendSerializtionError;
+import com.google.cloud.bigquery.storage.v1.Exceptions.AppendSerializationError;
 import com.google.cloud.bigquery.storage.v1.TableFieldSchema.Mode;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.Descriptors.DescriptorValidationException;
@@ -52,7 +50,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
-import java.util.logging.Logger;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.After;
@@ -67,13 +64,10 @@ import org.threeten.bp.LocalTime;
 
 @RunWith(JUnit4.class)
 public class JsonStreamWriterTest {
-  private static final Logger LOG = Logger.getLogger(JsonStreamWriterTest.class.getName());
-  private static int NUMERIC_SCALE = 9;
+  private static final int NUMERIC_SCALE = 9;
   private static final String TEST_STREAM = "projects/p/datasets/d/tables/t/streams/_default";
   private static final String TEST_STREAM_2 = "projects/p/datasets/d2/tables/t2/streams/_default";
   private static final String TEST_TABLE = "projects/p/datasets/d/tables/t";
-  private static final ExecutorProvider SINGLE_THREAD_EXECUTOR =
-      InstantiatingExecutorProvider.newBuilder().setExecutorThreadCount(1).build();
   private static LocalChannelProvider channelProvider;
   private FakeScheduledExecutorService fakeExecutor;
   private FakeBigQueryWrite testBigQueryWrite;
@@ -136,8 +130,7 @@ public class JsonStreamWriterTest {
   public void setUp() throws Exception {
     testBigQueryWrite = new FakeBigQueryWrite();
     serviceHelper =
-        new MockServiceHelper(
-            UUID.randomUUID().toString(), Arrays.<MockGrpcService>asList(testBigQueryWrite));
+        new MockServiceHelper(UUID.randomUUID().toString(), Arrays.asList(testBigQueryWrite));
     serviceHelper.start();
     channelProvider = serviceHelper.createChannelProvider();
     fakeExecutor = new FakeScheduledExecutorService();
@@ -638,7 +631,7 @@ public class JsonStreamWriterTest {
   }
 
   @Test
-  public void testCreateDefaultStreamWrongLocation() throws Exception {
+  public void testCreateDefaultStreamWrongLocation() {
     TableSchema tableSchema =
         TableSchema.newBuilder().addFields(0, TEST_INT).addFields(1, TEST_STRING).build();
     testBigQueryWrite.addResponse(
@@ -1096,9 +1089,9 @@ public class JsonStreamWriterTest {
       try {
         ApiFuture<AppendRowsResponse> appendFuture = writer.append(jsonArr);
         Assert.fail("expected ExecutionException");
-      } catch (AppendSerializtionError ex) {
+      } catch (AppendSerializationError ex) {
         assertEquals(
-            "JSONObject has fields unknown to BigQuery: root.test_unknown.",
+            "The source object has fields unknown to BigQuery: root.test_unknown.",
             ex.getRowIndexToErrorMessage().get(1));
         assertEquals(TEST_STREAM, ex.getStreamName());
       }
@@ -1188,7 +1181,7 @@ public class JsonStreamWriterTest {
   }
 
   @Test
-  public void testMultipleAppendSerializtionErrors()
+  public void testMultipleAppendSerializationErrors()
       throws DescriptorValidationException, IOException, InterruptedException {
     FooType expectedProto = FooType.newBuilder().setFoo("allen").build();
     JSONObject foo = new JSONObject();
@@ -1213,13 +1206,13 @@ public class JsonStreamWriterTest {
         getTestJsonStreamWriterBuilder(TEST_STREAM, TABLE_SCHEMA).build()) {
       try {
         ApiFuture<AppendRowsResponse> appendFuture = writer.append(jsonArr);
-        Assert.fail("expected AppendSerializtionError");
-      } catch (AppendSerializtionError appendSerializtionError) {
+        Assert.fail("expected AppendSerializationError");
+      } catch (AppendSerializationError appendSerializationError) {
         Map<Integer, String> rowIndexToErrorMessage =
-            appendSerializtionError.getRowIndexToErrorMessage();
+            appendSerializationError.getRowIndexToErrorMessage();
         assertEquals(2, rowIndexToErrorMessage.size());
         assertEquals(
-            "JSONObject has fields unknown to BigQuery: root.not_foo.",
+            "The source object has fields unknown to BigQuery: root.not_foo.",
             rowIndexToErrorMessage.get(0));
         assertEquals(
             "Field root.foo failed to convert to STRING. Error: JSONObject does not have a string field at root.foo.",
@@ -1253,10 +1246,10 @@ public class JsonStreamWriterTest {
         getTestJsonStreamWriterBuilder(TEST_STREAM, TABLE_SCHEMA).build()) {
       try {
         ApiFuture<AppendRowsResponse> appendFuture = writer.append(jsonArr);
-        Assert.fail("expected AppendSerializtionError");
-      } catch (AppendSerializtionError appendSerializtionError) {
+        Assert.fail("expected AppendSerializationError");
+      } catch (AppendSerializationError appendSerializationError) {
         Map<Integer, String> rowIndexToErrorMessage =
-            appendSerializtionError.getRowIndexToErrorMessage();
+            appendSerializationError.getRowIndexToErrorMessage();
         assertEquals(1, rowIndexToErrorMessage.size());
         assertTrue(
             rowIndexToErrorMessage
@@ -1310,7 +1303,7 @@ public class JsonStreamWriterTest {
     try (JsonStreamWriter writer =
         getTestJsonStreamWriterBuilder(TEST_STREAM, tableSchema).setTraceId("test:empty").build()) {
 
-      Map<String, AppendRowsRequest.MissingValueInterpretation> missingValueMap = new HashMap();
+      Map<String, AppendRowsRequest.MissingValueInterpretation> missingValueMap = new HashMap<>();
       missingValueMap.put("col1", AppendRowsRequest.MissingValueInterpretation.NULL_VALUE);
       missingValueMap.put("col3", AppendRowsRequest.MissingValueInterpretation.DEFAULT_VALUE);
       writer.setMissingValueInterpretationMap(missingValueMap);